# .stryker-tmp/sandbox-bCteHS/docs/api-contract.yaml
openapi: 3.0.3
info:
  title: VibeX API Contract
  description: |
    VibeX 全栈项目 API 契约文档
    
    ## 版本历史
    - v1.1.0 (2026-03-28): 从 14 端点扩展至 90+ 端点，覆盖全部前端实际调用
    - v1.0.0 (初始): 基础 14 端点定义
    
    ## 路径约定
    - 所有路径相对于 `/api` 前缀（如 `/auth/login` → `/api/auth/login`）
    - 服务器基础路径: `/api`
    - 统一无 v1 前缀（v1 双写路由已废弃，标注 `deprecated: true`）
    
    ## 路由分类
    - **前端调用**: 前端 services/api/ 模块中实际调用
    - **后端独有**: `x-backend-only: true`，前端未调用但后端已实现
    - **废弃路由**: `deprecated: true`，v1 双写兼容层
    
    ## 16 个 Tag 分组
    | Tag | 前缀 | 说明 |
    |-----|------|------|
    | Auth | /auth | 认证 |
    | Users | /users | 用户 |
    | Projects | /projects | 项目管理 |
    | Requirements | /requirements | 需求 |
    | Clarifications | /clarifications | 澄清问答 |
    | DomainEntities | /domain-entities, /domains | 域实体 |
    | EntityRelations | /entity-relations | 实体关系 |
    | DDD | /ddd | DDD 限界上下文/域模型/业务流程 |
    | Design | /clarify, /domain, /flow, /pages, /prototype, /design | 设计流程 |
    | Flows | /flows | 流程图 |
    | PrototypeSnapshots | /prototype-snapshots | 原型快照 |
    | Agents | /agents | Agent |
    | Pages | /pages | 页面 |
    | Chat | /chat | AI 对话 |
    | Plan | /plan | 计划分析 |
    | Diagnosis | /diagnosis | 需求诊断 |
    | BackendOnly | 各类后端独有路由 | 未被前端调用的后端路由 |
    | Deprecated | /api/v1/* | v1 兼容层（废弃） |
  version: 1.1.0
  contact:
    name: VibeX Team
servers:
  - url: /api
    description: API 基础路径
...
paths:
...
  /ui-generation:
    post:
      tags: [BackendOnly]
      summary: UI 生成（流式）
      operationId: uiGeneration
      x-backend-only: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [spec]
              properties:
                spec:
                  type: string
                theme:
                  type: object
      responses:
        '200':
          description: 流式响应
          content:
            text/event-stream:
              schema:
                type: string
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...